pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "sova_relay"  # adjust as desired
version = "0.1.0"
authors = [
    { name = "Your Name", email = "you@example.com" }
]
description = "Adapter between a PWM relay and a virtual Sokil relay on Raspberry Pi."
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.8"
dependencies = [
    "RPi.GPIO",
    "rpi-lgpio"
]

[project.urls]
"Homepage" = "https://example.com/"

[tool.setuptools.packages.find]
where = ["src"]
